requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
langchain-groq>=0.0.1
faiss-cpu>=1.7.4
PyMuPDF>=1.23.8
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
pypdf>=3.17.1
tiktoken>=0.5.1
